+++ core/src/main/scala/com/stratio/sparta/core/utils/Utils.scala
@@ -34,4 +34,10 @@ object Utils extends SLF4JLogging {
     }
   }
 
+  def using[A <: {def close() : Unit}, B](resource: A)(f: A => B): B =
+    try {
+      f(resource)
+    } finally {
+      resource.close()
+    }
 }